+++ app/maintainers/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MaintainersConfig(AppConfig):
+    name = 'maintainers'
