+++ app/maintainers/serializers.py
@@ -1,0 +1,1 @@
+from rest_framework import serializers
