+++ app/urls.py
@@ -20,12 +20,10 @@ from ports import views
 urlpatterns = [
     path('', views.index, name='home'),
     path('statistics/', include('stats.urls')),
-    url(r'^maintainer/github/(?P<github_handle>[-a-zA-Z0-9_.]+)/$', views.maintainer_detail_github, name='maintainer_detail_github'),
-    url(r'^maintainer/email/(?P<name>[-a-zA-Z0-9_.]+)__(?P<domain>[-a-zA-Z0-9_.]+)/$', views.maintainer_detail_email, name='maintainer_detail_email'),
+    path('maintainers/', include('maintainers.urls')),
     path('port/', include('port_detail.urls')),
     path('ports/', views.index, name='ports-index'),
     path('ports/search/', views.search, name='ports_search'),
-    path('ports/filter/maintainer/', views.search_ports_in_maintainer, name='search_ports_in_maintainer'),
     path('ports/filter/category/', views.search_ports_in_category, name='search_ports_in_category'),
     path('ports/filter/variant/', views.search_ports_in_variant, name='search_ports_in_variant'),
     path('ports/load_tickets/', views.tickets, name='trac_tickets'),
